fix bugs; merge files
