+++ input/parser.go
@@ -2,7 +2,9 @@ package input
 
 import (
 	"encoding/json"
+	"strconv"
 	"strings"
+	"time"
 
 	"github.com/ekanite/ekanite/input/syslog"
 	"github.com/ekanite/ekanite/input/syslog/rfc3164"
@@ -10,6 +12,7 @@ import (
 )
 
 var (
+	ok                  bool
 	FORMATS_BY_STANDARD = []string{"rfc3164", "rfc5424", "ecma404"}
 	FORMATS_BY_NAME     = []string{"syslog-bsd", "syslog", "json"}
 )
@@ -18,6 +21,12 @@ type Input struct {
 	Format   string
 	Parsed   map[string]interface{}
 	Unparsed []byte
+}
+
+type Timestamp struct {
+	Unparsed []string
+	Fromated []int64
+	Parsed   string
 }
 
 func NewParser(f string) Input {
@@ -132,15 +141,13 @@ func (i *Input) parseJson() bool {
 	}
 
 	stats.Add("ecma404Parsed", 1)
-	return true
+	return i.parseTimestamp()
 
 }
 
 func (i *Input) parseTimestamp() bool {
 
-	_, ok := i.Parsed["timestamp"]
-
-	if !ok {
+	if _, ok := i.Parsed["timestamp"]; !ok {
 
 		stats.Add("ecma404MissingTimestamp", 1)
 		return false
@@ -156,5 +163,61 @@ func (i *Input) parseTimestamp() bool {
 	}
 
 	return true
+
+}
+
+var err error
+
+func NewTiemstamp() *Timestamp {
+
+	return &Timestamp{}
+
+}
+
+func (t *Timestamp) Parse(ts string) (bool, string) {
+
+	t.Unparsed = strings.Split(ts, ".")
+
+	if ok := t.format(); !ok {
+
+		return false, t.Parsed
+
+	}
+
+	t.Parsed = time.Unix(t.Fromated[0], t.Fromated[1]).Format(time.RFC3339)
+
+	return true, t.Parsed
+
+}
+
+func (t *Timestamp) format() bool {
+
+	for _, ts := range t.Unparsed {
+
+		p, err := strconv.ParseInt(ts, 10, 64)
+
+		if err != nil {
+
+			stats.Add("UnformatedTimestampPart", 1)
+			return false
+
+		}
+
+		t.Fromated = append(t.Fromated, p)
+
+	}
+
+	t.ensureLength()
+	return true
+
+}
+
+func (t *Timestamp) ensureLength() {
+
+	if len(t.Fromated) == 1 {
+
+		t.Fromated = append(t.Fromated, 0)
+
+	}
 
 }
